// generated/loopringExplorer.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
};

export type Account = {
  /** Address linked to the account ID */
  address: Scalars['Bytes'];
  /** List of all the balances for each specific token the account has interacted with */
  balances: Array<AccountTokenBalance>;
  /** L2 transaction internalID where the account was first created and linked to an address. Useful for sorting and filtering purposes */
  createdAt: Scalars['BigDecimal'];
  /** L2 transaction where the account was first created and linked to an address */
  createdAtTransaction: Transaction;
  /** Internal ID used in the L2 transactions. Users will be of type User and have an ID > 10000. Pools will be of type Pool and have an id < 10000. ProtocolAccount is a special account with ID 0 which handles all token balances generated by protocol fees. */
  id: Scalars['ID'];
  /** Same as ID but expressed as a BigInt for sorting purposes */
  internalID: Scalars['BigInt'];
  /** L2 transaction internalID that last updated the account entity. Useful for sorting and filtering purposes */
  lastUpdatedAt: Scalars['BigDecimal'];
  /** L2 transaction that last updated the account entity */
  lastUpdatedAtTransaction: Transaction;
  /** List of all the AccountNFTSlot entities that this account has. Those slots can be empty but will only exist if they held an NFT at some point. */
  slots: Array<AccountNftSlot>;
  /** L2 transactions that involved this account */
  transactions: Array<Transaction>;
};
// ...
export type AccountNftSlot = {
  __typename?: 'AccountNFTSlot';
  /** Owner of this slot */
  account: Account;
  /** Amount of copies of this NonFungibleToken saved in this slot. */
  balance: Scalars['BigInt'];
  /** L2 transaction internalID where the slot was first created. Useful for sorting and filtering purposes */
  createdAt: Scalars['BigDecimal'];
  /** L2 transaction where the slot was first created */
  createdAtTransaction: Transaction;
  /** ID is compounded from the account ID and the token ID used as an NFT slot. */
  id: Scalars['ID'];
  /** L2 transaction internalID that last updated the slot entity. Useful for sorting and filtering purposes */
  lastUpdatedAt: Scalars['BigDecimal'];
  /** L2 transaction that last updated the slot entity */
  lastUpdatedAtTransaction: Transaction;
  /** NonFungibleToken saved in this slot. Null if empty */
  nft?: Maybe<NonFungibleToken>;
  /** List of all the transactions that this particular NFT was a part of */
  transactions: Array<TransactionNft>;
};
// ...
export type AccountTokenBalance = {
  __typename?: 'AccountTokenBalance';
  /** Link to the Account entity. Can be a Pool or User entity */
  account: Account;
  balance: Scalars['BigInt'];
  dailyData: Array<AccountTokenBalanceDailyData>;
  /** ID is recreated following this pattern: <ACCOUNT ID>-<TOKEN ID>. Example: 10001-1 */
  id: Scalars['ID'];
  /** Link to the Token entity. Holds all details of the token itself. */
  token: Token;
  transactions: Array<Transaction>;
  weeklyData: Array<AccountTokenBalanceWeeklyData>;
};
// ...
export type AccountTokenBalanceDailyData = {
  __typename?: 'AccountTokenBalanceDailyData';
  /** Link to the Account entity. Can be a Pool or User entity */
  account: Account;
  /** Link to the original AccountTokenBalance entity. */
  accountTokenBalance: AccountTokenBalance;
  /** Latest value for balance for this AccountTokenBalance entity on this specific day. Coincides with balanceClose */
  balance: Scalars['BigInt'];
  /** Last update of balance for this AccountTokenBalance entity on this specific day */
  balanceClose: Scalars['BigInt'];
  /** Highest balance for this AccountTokenBalance entity on this specific day */
  balanceHigh: Scalars['BigInt'];
  /** Lowest balance for this AccountTokenBalance entity on this specific day */
  balanceLow: Scalars['BigInt'];
  /** Balance for this AccountTokenBalance entity when it was created. */
  balanceOpen: Scalars['BigInt'];
  /** Timestamp of the end of the day. Timezone -> UTC */
  dayEnd: Scalars['BigInt'];
  /** Day number is the amount of days since the start block of Loopring 3.6 (block 11149814) */
  dayNumber: Scalars['Int'];
  /** Timestamp of the start of the day. Timezone -> UTC */
  dayStart: Scalars['BigInt'];
  /** The ID follows this pattern: <ACCOUNT TOKEN BALANCE ID>-<DAY NUMBER>. */
  id: Scalars['ID'];
  /** Link to the Token entity. Holds all details of the token itself. */
  token: Token;
};
// ...
export type AccountTokenBalanceWeeklyData = {
  __typename?: 'AccountTokenBalanceWeeklyData';
  /** Link to the Account entity. Can be a Pool or User entity */
  account: Account;
  /** Link to the original AccountTokenBalance entity. */
  accountTokenBalance: AccountTokenBalance;
  /** Latest value for balance for this AccountTokenBalance entity on this specific week. Coincides with balanceClose */
  balance: Scalars['BigInt'];
  /** Last update of balance for this AccountTokenBalance entity on this specific week */
  balanceClose: Scalars['BigInt'];
  /** Highest balance for this AccountTokenBalance entity on this specific week */
  balanceHigh: Scalars['BigInt'];
  /** Lowest balance for this AccountTokenBalance entity on this specific week */
  balanceLow: Scalars['BigInt'];
  /** Balance for this AccountTokenBalance entity when it was created. */
  balanceOpen: Scalars['BigInt'];
  /** The ID follows this pattern: <ACCOUNT TOKEN BALANCE ID>-<WEEK NUMBER>. */
  id: Scalars['ID'];
  /** Link to the Token entity. Holds all details of the token itself. */
  token: Token;
  /** Timestamp of the end of the week. Timezone -> UTC */
  weekEnd: Scalars['BigInt'];
  /** Week number is the amount of weeks since the start block of Loopring 3.6 (block 11149814) */
  weekNumber: Scalars['Int'];
  /** Timestamp of the start of the week. Timezone -> UTC */
  weekStart: Scalars['BigInt'];
};
// ...
export type Exchange = {
  __typename?: 'Exchange';
  /** Address of the exchange implementation */
  id: Scalars['ID'];
  /** Proxy link. Used to generate derived array of implementations. */
  proxy: Proxy;
  /** List of all tokens supported during the time this Exchange was the current implementation */
  tokens: Array<Token>;
};
// ...
export enum OrderDirection {
  Asc = 'asc',
  Desc = 'desc',
}
// ...
export type Query = {
  __typename?: 'Query';
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  account?: Maybe<Account>;
  accountNFTSlot?: Maybe<AccountNftSlot>;
  accountNFTSlots: Array<AccountNftSlot>;
  accountTokenBalance?: Maybe<AccountTokenBalance>;
  accountTokenBalanceDailyData?: Maybe<AccountTokenBalanceDailyData>;
  accountTokenBalanceDailyDatas: Array<AccountTokenBalanceDailyData>;
  accountTokenBalanceWeeklyData?: Maybe<AccountTokenBalanceWeeklyData>;
  accountTokenBalanceWeeklyDatas: Array<AccountTokenBalanceWeeklyData>;
  accountTokenBalances: Array<AccountTokenBalance>;
  accountUpdate?: Maybe<AccountUpdate>;
  accountUpdates: Array<AccountUpdate>;
  accounts: Array<Account>;
  add?: Maybe<Add>;
  adds: Array<Add>;
  ammUpdate?: Maybe<AmmUpdate>;
  ammUpdates: Array<AmmUpdate>;
  block?: Maybe<Block>;
  blocks: Array<Block>;
  dataNFT?: Maybe<DataNft>;
  dataNFTs: Array<DataNft>;
  deposit?: Maybe<Deposit>;
  deposits: Array<Deposit>;
  exchange?: Maybe<Exchange>;
  exchanges: Array<Exchange>;
  mintNFT?: Maybe<MintNft>;
  mintNFTs: Array<MintNft>;
  nonFungibleToken?: Maybe<NonFungibleToken>;
  nonFungibleTokens: Array<NonFungibleToken>;
  orderbookTrade?: Maybe<OrderbookTrade>;
  orderbookTrades: Array<OrderbookTrade>;
  pair?: Maybe<Pair>;
  pairDailyData?: Maybe<PairDailyData>;
  pairDailyDatas: Array<PairDailyData>;
  pairWeeklyData?: Maybe<PairWeeklyData>;
  pairWeeklyDatas: Array<PairWeeklyData>;
  pairs: Array<Pair>;
  pool?: Maybe<Pool>;
  pools: Array<Pool>;
  protocolAccount?: Maybe<ProtocolAccount>;
  protocolAccounts: Array<ProtocolAccount>;
  proxies: Array<Proxy>;
  proxy?: Maybe<Proxy>;
  remove?: Maybe<Remove>;
  removes: Array<Remove>;
  signatureVerification?: Maybe<SignatureVerification>;
  signatureVerifications: Array<SignatureVerification>;
  swap?: Maybe<Swap>;
  swapNFT?: Maybe<SwapNft>;
  swapNFTs: Array<SwapNft>;
  swaps: Array<Swap>;
  token?: Maybe<Token>;
  tokenDailyData?: Maybe<TokenDailyData>;
  tokenDailyDatas: Array<TokenDailyData>;
  tokenWeeklyData?: Maybe<TokenWeeklyData>;
  tokenWeeklyDatas: Array<TokenWeeklyData>;
  tokens: Array<Token>;
  tradeNFT?: Maybe<TradeNft>;
  tradeNFTs: Array<TradeNft>;
  transaction?: Maybe<Transaction>;
  transactionNFT?: Maybe<TransactionNft>;
  transactionNFTs: Array<TransactionNft>;
  transactions: Array<Transaction>;
  transfer?: Maybe<Transfer>;
  transferNFT?: Maybe<TransferNft>;
  transferNFTs: Array<TransferNft>;
  transfers: Array<Transfer>;
  user?: Maybe<User>;
  users: Array<User>;
  withdrawal?: Maybe<Withdrawal>;
  withdrawalNFT?: Maybe<WithdrawalNft>;
  withdrawalNFTs: Array<WithdrawalNft>;
  withdrawals: Array<Withdrawal>;
};
// ...
export type Token = {
  __typename?: 'Token';
  /** Address of the token within the Ethereum network */
  address: Scalars['Bytes'];
  dailyEntities: Array<TokenDailyData>;
  decimals: Scalars['Int'];
  /** Exchange active when the token was registered */
  exchange: Exchange;
  /** Internal ID of the token within Loopring */
  id: Scalars['ID'];
  /** Internal ID of the token within loopring expressed as a number for ordering */
  internalID: Scalars['BigInt'];
  name: Scalars['String'];
  symbol: Scalars['String'];
  /** Total traded volume for this token across all token pairs. Denominated in it's own currency */
  tradedVolume: Scalars['BigInt'];
  /** Total traded volume for this token across all token pairs. Denominated in it's own currency */
  tradedVolumeOrderbook: Scalars['BigInt'];
  /** Total traded volume for this token across all token pairs. Denominated in it's own currency */
  tradedVolumeSwap: Scalars['BigInt'];
  weeklyEntities: Array<TokenWeeklyData>;
};
// ...
export type TokenDailyData = {
  __typename?: 'TokenDailyData';
  /** Timestamp of the end of the day. Timezone -> UTC */
  dayEnd: Scalars['BigInt'];
  /** Day number is the amount of days since the start block of Loopring 3.6 (block 11149814) */
  dayNumber: Scalars['Int'];
  /** Timestamp of the start of the day. Timezone -> UTC */
  dayStart: Scalars['BigInt'];
  /** The ID follows this pattern: <TOKEN ID>-<DAY NUMBER>. */
  id: Scalars['ID'];
  token: Token;
  /** Volume traded for the token. Includes both Swaps and Orderbook trades */
  tradedVolume: Scalars['BigInt'];
  /** Volume traded for the token exclusively on Orderbook trades */
  tradedVolumeOrderbook: Scalars['BigInt'];
  /** Volume traded for the token exclusively on Swaps */
  tradedVolumeSwap: Scalars['BigInt'];
};
// ...
export type Transaction = {
  accounts?: Maybe<Array<Account>>;
  /** Link to the Block entity where this Transaction took place */
  block: Block;
  /** Hex string representation of the encoded L2 data for this Transaction */
  data: Scalars['String'];
  /** All transactions IDs follow the same pattern: <BLOCK ID>-<TX INDEX> */
  id: Scalars['ID'];
  /** ID represented as a BigDecimal for sorting purposes */
  internalID: Scalars['BigDecimal'];
  tokenBalances?: Maybe<Array<AccountTokenBalance>>;
  /** Explicit copy of __typename to make it usable when filtering */
  typename: TransactionType;
};
// ...
export type AccountTokenBalancesQuery = {
  __typename?: 'Query';
  accountTokenBalances: Array<{
    __typename: 'AccountTokenBalance';
    id: string;
    balance: any;
    token: { __typename?: 'Token'; id: string; name: string; symbol: string; decimals: number; address: any };
  }>;
};